Bugfix: prevent that tags are deleted.
The following sequence does not work:

$fwObject = Get-AzFirewall -Name $firewallName
Set-AzResource -Tag @{FirewallInfo=($fwEntity | ConvertTo-Json)} -ResourceId $fwObject.Id -Force
Set-AzFirewall -AzureFirewall $fwObject

when performing line 3, we use the object as declared in line 1, which is without the new tag. So line 3 reverts the changes done in line 2. 

Possible solutions:
a) switch line 2 and 3 (proposed solution in the pull request)
b) replace line 2 so the code modifies the $fwObject.